+++ src/main/java/de/screenflow/frankenstein/vf/local/BWConfigController.java
@@ -1,5 +1,0 @@
-package de.screenflow.frankenstein.vf.local;
-
-public class BWConfigController {
-
-}