+++ RELEASE-NOTES.md
@@ -6,5 +6,6 @@
 ## Version 0.2 (currently working on)
 - Cam Stream Input (no audio)
 - Network Stream Input (opencv or vlc/file)
-- Video Segment Filters
+- Segment Filters (Black and White, Stereo Distance)
+